+++ shared/js/ui/frames/side/ChannelConversationController.ts
@@ -3,7 +3,7 @@ import {EventHandler} from "tc-events";
 import {LogCategory, logError} from "../../../log";
 import {tr} from "../../../i18n/localize";
 import {AbstractConversationUiEvents} from "./AbstractConversationDefinitions";
-import {AbstractConversationController} from "./AbstractConversationController";
+import {AbstractConversationController, SelectedConversation} from "./AbstractConversationController";
 import {
     ChannelConversation,
     ChannelConversationEvents,
@@ -11,6 +11,7 @@ import {
     ChannelConversationManagerEvents
 } from "tc-shared/conversations/ChannelConversationManager";
 import {ChannelConversationUiEvents} from "tc-shared/ui/frames/side/ChannelConversationDefinitions";
+import {spawnModalChannelChat} from "tc-shared/ui/modal/channel-chat/Controller";
 
 export class ChannelConversationController extends AbstractConversationController<
     ChannelConversationUiEvents,
@@ -26,17 +27,15 @@ export class ChannelConversationController extends AbstractConversationControlle
         super();
         this.connectionListener = [];
 
-        /*
-        spawnExternalModal("conversation", this.uiEvents, {
-            handlerId: this.connection.handlerId,
-            noFirstMessageOverlay: false,
-            messagesDeletable: true
-        }).open().then(() => {
-            console.error("Opened");
-        });
-        */
-
         this.uiEvents.registerHandler(this, true);
+        this.uiEvents.on("action_popout_chat", () => {
+            const conversation = this.getCurrentConversation();
+            if(!conversation) {
+                return;
+            }
+
+            spawnModalChannelChat(this.connection, conversation);
+        });
     }
 
     destroy() {
@@ -59,9 +58,14 @@ export class ChannelConversationController extends AbstractConversationControlle
         if(connection) {
             /* FIXME: Update cross channel talk state! */
             this.setConversationManager(connection.getChannelConversations());
+            this.setSelectedConversation("conversation-manager-selected");
         } else {
             this.setConversationManager(undefined);
         }
+    }
+
+    setSelectedConversation(conversation: SelectedConversation<ChannelConversation>) {
+        super.setSelectedConversation(conversation);
     }
 
     @EventHandler<AbstractConversationUiEvents>("action_delete_message")
@@ -75,15 +79,17 @@ export class ChannelConversationController extends AbstractConversationControlle
         conversation.deleteMessage(event.uniqueId);
     }
 
-    protected registerConversationEvents(conversation: ChannelConversation) {
-        super.registerConversationEvents(conversation);
+    protected registerConversationEvents(conversation: ChannelConversation): (() => void)[] {
+        const events = super.registerConversationEvents(conversation);
 
-        this.currentSelectedListener.push(conversation.events.on("notify_messages_deleted", event => {
+        events.push(conversation.events.on("notify_messages_deleted", event => {
             this.uiEvents.fire_react("notify_chat_message_delete", { messageIds: event.messages, chatId: conversation.getChatId() });
         }));
 
-        this.currentSelectedListener.push(conversation.events.on("notify_conversation_mode_changed", () => {
+        events.push(conversation.events.on("notify_conversation_mode_changed", () => {
             this.reportStateToUI(conversation);
         }));
+
+        return events;
     }
 }