+++ shared/js/ui/frames/side/ChannelConversationDefinitions.ts
@@ -1,3 +1,3 @@
 import {AbstractConversationUiEvents} from "tc-shared/ui/frames/side/AbstractConversationDefinitions";
 
-export interface ChannelConversationUiEvents extends AbstractConversationUiEvents {}
+export interface ChannelConversationUiEvents extends AbstractConversationUiEvents { }